Сделала из одного адаптера - два разных
